move db types into their own file
